Kill off redundant and incorrect sizing definition.
It seems that in investigating the fencepost issue (issue59), it was
revealed that the 255 value of LARGEST_ID was well beyond what was
ever possible to be referenced.

As actual value was limited to the POWER_LARGEST definition in
practice, I have LARGEST_ID use POWER_LARGEST to define itself.